+++ Changes.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## v0.2.1
+* Added a pseudo real-time clock block for real-time simulation
+* Added another demo: small campus with central chiller plant
+* Added demonstration plotting scripts to several demos
+* Added basic test script for EnergyPlus / MLE+
+* Updated documentation to indicate deprecation of DataBus functions
+* Changed license to LGPL
+
 ## v0.2
 * Added single-phase constant power electrical source/load block
 * Added time alignment options to PVWatts cosimulation block